chore: rename old utils.go that contained cli styling to style.go
Move general utility function ShuffleFrom to utils.go

Changes are done for clarity
